refactor leaflet component. add population density layer. break volcano pic
